+++ CRM/src/main/java/com/sc/mapper/SysEmpuserMapper.java
@@ -2,7 +2,6 @@ package com.sc.mapper;
 
 import com.sc.bean.SysEmpuser;
 import com.sc.bean.SysEmpuserExample;
-import java.math.BigDecimal;
 import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
@@ -11,7 +10,7 @@ public interface SysEmpuserMapper {
 
     int deleteByExample(SysEmpuserExample example);
 
-    int deleteByPrimaryKey(BigDecimal empId);
+    int deleteByPrimaryKey(Long empId);
 
     int insert(SysEmpuser record);
 
@@ -19,7 +18,7 @@ public interface SysEmpuserMapper {
 
     List<SysEmpuser> selectByExample(SysEmpuserExample example);
 
-    SysEmpuser selectByPrimaryKey(BigDecimal empId);
+    SysEmpuser selectByPrimaryKey(Long empId);
 
     int updateByExampleSelective(@Param("record") SysEmpuser record, @Param("example") SysEmpuserExample example);
